Adds true support for [Nvidia's cuDNN](https://developer.nvidia.com/cudnn), and reorganized repo to be more user friendly. cuDNN speedup is more noticeable when backprop is the bottleneck.

+++ Distributions/Categorical.h
@@ -22,7 +22,7 @@ public:
     torch::Tensor enumerate_support();
 
     torch::Tensor m_logits, m_probs;
-    int m_num_events;
+    int64_t m_num_events;
 
     std::shared_ptr<torch::Device> m_device;
